Added listing pokemon teams by type.
After adding more Pokemon, the time it takes to process each team grows. On top of this, there are too many teams for an individual to sieve through.
To fix this issue, we added functionality to list the Pokemon teams by type, as the main functionality that matters is team maching by Pokemon type pairs caught by each player.
We then list each Pokemon for their teams that are in that pairing, making it easier to review.

They are also ordered in the same pairs, meaning the nth Pokemon in the mth type listing for player one will also match with the same (n, m) index for player two.