Added death check for Pokemon pair in csv
Optionally, you can add "Dead" or some other flag to the end of a line
to signify the pokemon pair is dead. This allows easy tracking of
previously caught Pokemon. You can also add "Alive" (case-insensitive)
to the end of the line for documentation, but assumes the pair is alive
by default. You can also add any other comments to have the pokemon pair
be ignored, such as "Missed" for missed captures.